완벽히 라이트닝 끝

+++ Assets/LightningPoTap.cs
@@ -20,61 +20,70 @@ public class LightningPoTap : PoTapBase
     }
     private void FireLightning(){
         if(targetTransform!=null&&gunDeley>=myGunDeley){
+            FindTarger(transform.position, ref targetTransform);
+            if(targetTransform == null)return;
             gunDeley=0f;
             boundCount=0;
             poss.RemoveRange(0,poss.Count);
             poss.Add(shootingPos);
+            poss.Add(targetTransform);
             NextTarget();
 
-            for(int i=0;i<poss.Count;i++)
-                Debug.Log(poss[i]);
+            // for(int i=0;i<poss.Count;i++)
+            //     Debug.Log(poss[i]);
             AllPoolManager.Instance.GetObjPos(3,shootingPos).gameObject.SetActive(true);
             myParticleSystem.Play();
         }
     }
-    protected override private void FindTarger(Vector3 pos, ref Transform _targetTransform){
-        Transform near=null;
+    private void FindLightningTarger(Vector2 pos){
+        if(boundCount>bound)return;
         float nearRange, nearRange2=range;
+        Transform near=null;
+        bool hi=false;
         Collider2D[] cols = Physics2D.OverlapCircleAll(pos, range, whatLayerMask);
         if(cols.Length==0)return;
         foreach (Collider2D col in cols)
         {
-            foreach(Transform tran in poss){
-                
+            hi=false;
+            Debug.Log("심판시작 "+col);
+            foreach (Transform tran in poss)
+            {
                 if(tran == col.transform){
-                    Debug.Log(tran);
-                    Debug.Log("중복");
-                    continue;
+                    Debug.Log("넌 겹친다 "+col);
+                    hi=true;
+                    break;
                 }
             }
-            nearRange=Vector3.Distance(playerPos.position , col.transform.position);
-            if(nearRange2>nearRange){
+            if(hi)continue;
+            nearRange=Vector2.Distance(pos , col.transform.position);                     
+            if(nearRange2>=nearRange){
                 nearRange2 = nearRange; 
                 near = col.transform;
+                Debug.Log("통과"+nearRange);
+            }else{
+                Debug.Log("탈락"+nearRange);
             }
+            
         }
-        if(near==null)return;
-        
-        _targetTransform = near;
-    }
-    private void NextTarget(){
-        newTarget=null;
-        Debug.Log(boundCount);
-        FindTarger(poss[boundCount].position,ref newTarget);
-        poss.Add(newTarget);
-        if(bound>boundCount||newTarget==null){
-            boundCount++;
-            NextTarget();
-        }
-        else{
+        if(near == null){
+            Debug.Log("대상을 찾을수 없음.");
             return;
         }
+        poss.Add(near);
+        Debug.Log(boundCount);
+        boundCount++;
+        FindLightningTarger(near.position);
+    }
+    private void NextTarget(){
+        boundCount=0;
+        FindLightningTarger(poss[1].position);
     }
     private void SetLinePos(){
         for(int i=0;i<poss.Count;i++){
             if(poss[i]==null)return;
             lineRenderer.positionCount = i+1;
-            lineRenderer.SetPosition(i,poss[i].position);
+            Vector2 pos = poss[i].position;
+            lineRenderer.SetPosition(i,pos);
         }
         
     }
